Rename network::Client to network::NetworkClient
