+++ fastpay_core/src/lib.rs
@@ -1,12 +1,5 @@
 // Copyright (c) Facebook, Inc. and its affiliates.
 // SPDX-License-Identifier: Apache-2.0
-#![warn(
-    future_incompatible,
-    nonstandard_style,
-    rust_2018_idioms,
-    rust_2021_compatibility
-)]
-#![deny(warnings)]
 
 pub mod authority;
 pub mod client;
